+++ src/checker/inference.rs
@@ -687,6 +687,7 @@ impl Checker {
                 let callable = self.infer(callee)?;
                 let target = self.indirect_callable_target(&callable);
                 let (ret, _, error) = self.infer_callable_ty(
+                    &expr.source_span(),
                     "<callable>",
                     callable.clone(),
                     param_args,